Move vector_scope->main_view_model and some rename
