+++ src-tauri/src/model.rs
@@ -1,5 +1,4 @@
 pub mod graph_plotter;
 pub mod mouse_info;
 pub mod screenshot_capture;
-pub mod vector_scope;
-pub mod worker_thread;
+pub mod worker_thread_base;
